Code refactor based on pull request - part 4 - Front-end Composition
