+++ interactive/app_utils.py
@@ -1,16 +1,18 @@
 import streamlit as st
 
 from PIL import Image
+from loguru import logger
 
-from common.consts import INTERACTIVE_ASSETS_DICT
 from interactive import (
     CATEGORY_DESCR,
     IMAGE_EXAMPLES,
 )
 from common import env_handler
+from interactive import GRID_NROW_NUMBER
 
 from metrics.consts import MetricCollections
 from metrics.core import MetricClient
+from common.consts import INTERACTIVE_ASSETS_DICT
 
 
 class ModuleManager:
@@ -18,9 +20,71 @@ class ModuleManager:
     List of components used for building the app.
     """
 
-    def __init__(self, states_manager) -> None:
-        self.states_manager = states_manager
+    def __init__(self) -> None:
         self.metric_client = MetricClient()
+
+    def widget_formatting(self) -> None:
+        with open(INTERACTIVE_ASSETS_DICT["app_style_file"], "r") as f:
+            st.markdown(f"<style>{f.read}</style>", unsafe_allow_html=True)
+
+    def initialize_states(self) -> None:
+        """
+        Initializes states within the application. Those are distinct for every user.
+        The "if" condition is necessary for the app to not reset it's whole session_state.
+        """
+        if "init" not in st.session_state:
+            st.session_state.init = True
+            st.session_state.category_desc_option = None
+            st.session_state.category_option = None
+            st.session_state.provisioning_options = None
+
+            # Option 1 States - Example Images
+            st.session_state.example_captions = None
+            st.session_state.example_imgs = None
+
+            # Option 2 States - Storage Images
+            st.session_state.pull_random_img_number = None
+            st.session_state.refresh_random_images = None
+            st.session_state.random_captions = None
+            st.session_state.random_imgs = None
+
+            # Option 3 States - Uploaded File Images
+            # Placeholder
+
+            # All Options States - set when an input image has been selected
+            st.session_state.show_input_img = None
+            st.session_state.selected_img = None
+
+            # Search and Output States - set when an input image has been selected and before "Find Similar Images" is run
+            st.session_state.similar_img_number = None
+            st.session_state.benchmark_similarity_value = None
+            st.session_state.grid_nrow_number = GRID_NROW_NUMBER
+
+            # Search Completed State - set after "Find Similar Images" is completed
+            st.session_state.similar_images_found = None
+
+    def reset_all_states_button(self) -> None:
+        """
+        Reset all application starting from category selection.
+        """
+        st.session_state.category_desc_option = None
+        st.session_state.category_option = None
+        st.session_state.provisioning_options = None
+        st.session_state.show_input_img = None
+        st.session_state.selected_img = None
+        st.session_state.benchmark_similarity_value = None
+        st.session_state.similar_img_number = None
+        st.session_state.similar_images_found = None
+
+    def reset_states_after_image_provisioning_list(self) -> None:
+        """
+        Reset all application states after radio selection for image provisioning.
+        """
+        st.session_state.show_input_img = None
+        st.session_state.selected_img = None
+        st.session_state.benchmark_similarity_value = None
+        st.session_state.similar_img_number = None
+        st.session_state.similar_images_found = None
 
     def create_header(self) -> None:
         """
@@ -42,7 +106,7 @@ class ModuleManager:
         """
         st.header("Input Options")  # header
         if st.button("Reset All"):  # reset all button
-            self.states_manager.reset_all_states_button()
+            self.reset_all_states_button()
         st.write("")
         st.markdown(
             f'<p class="big-font">Which category would you like to search from?</p>',
@@ -73,7 +137,7 @@ class ModuleManager:
         Resets state of previous provisioning selection and creates a category-specific list of image examples
         that a user can select from.
         """
-        self.states_manager.reset_states_after_image_provisioning_list()
+        self.reset_states_after_image_provisioning_list()
         st.session_state.example_captions = [
             s_img["label"] for s_img in IMAGE_EXAMPLES[st.session_state.category_option.value]
         ]  # get captions
@@ -97,7 +161,7 @@ class ModuleManager:
         image examples that a user can select from the list. Additionally, a button for re-running random selection is
         implemented together with the input option for the number of sampled images.
         """
-        self.states_manager.reset_states_after_image_provisioning_list()
+        self.reset_states_after_image_provisioning_list()
         st.session_state.pull_random_img_number = st.number_input(
             label=f"Choose random images from {st.session_state.category_option.value} category.",
             value=5,
@@ -130,7 +194,7 @@ class ModuleManager:
         Resets state of previous provisioning selection and provides upload button for a user.
         Any RGB image can be uploaded.
         """
-        self.states_manager.reset_states_after_image_provisioning_list()
+        self.reset_states_after_image_provisioning_list()
         st.markdown(f'<p class="big-font">Choose a file.</p>', unsafe_allow_html=True)
         byte_img = st.file_uploader("Upload an image from a local disk.")
         if byte_img:
@@ -201,4 +265,59 @@ class ModuleManager:
         )
 
         if st.button("Reset Images"):
-            self.states_manager.reset_all_states_button()
+            self.reset_all_states_button()
+
+
+    def run_app(self):
+        logger.info("Set main graphical options.")
+        st.set_page_config(layout="wide")
+        self.widget_formatting()
+
+        logger.info("Create a header with initial paragraph.")
+        self.create_header()
+
+        logger.info("Initialize states.")
+        self.initialize_states()
+
+        logger.info("Create Main Filters - till category search")
+        self.create_main_filters()
+
+        logger.info("Columnar split")
+        col_image_1, col_image_2 = st.columns([1, 1])
+
+        logger.info("Image Provisioning")
+        with col_image_1:
+            logger.info("Image Provisioning Type Selection")
+            provisioning_options = [
+                "Example List",
+                "Pull Randomly from the Storage",
+                "File Upload",
+            ]
+            self.create_image_provisioning_options(
+                provisioning_options=provisioning_options
+            )
+        with col_image_2:
+            logger.info(
+                "Image Selection - only if st.session_state.category_option of main filters was chosen beforehand."
+            )
+            if st.session_state.category_option:
+                if st.session_state.provisioning_options == provisioning_options[0]:
+                    self.create_image_provision_for_examples()
+                elif st.session_state.provisioning_options == provisioning_options[1]:
+                    self.create_image_provision_for_random_storage_pull()
+                elif st.session_state.provisioning_options == provisioning_options[2]:
+                    self.create_image_provision_for_manual_upload()
+
+        logger.info("Shows an Image selected in the provisioning stage")
+        if st.session_state.show_input_img:
+            st.header("Input Image")
+            input_img_placeholder = st.empty()
+            input_img_placeholder.image(st.session_state.selected_img)
+
+        if st.session_state.category_option and st.session_state.selected_img:
+            logger.info("Similarity Search Filters")
+            self.create_similarity_search_filters()
+
+            logger.info("Similarity Search Button")
+            if st.button("Find Similar Images"):
+                self.extract_similar_images()
